# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CineNova API
  version: 0.1.0
  summary: Rights-aware streaming platform API foundation.
  description: >-
    Initial REST API contract for mock licensed catalogue, search, playback session creation,
    downloads, and provider health. All sensitive provider calls remain server-side.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/v1/search:
    get:
      summary: Search the licensed catalogue
      operationId: searchCatalogue
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 120
        - name: region
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
            default: NG
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/Problem'
...
  /api/v1/playback/session:
    post:
      summary: Create a short-lived playback session
      operationId: createPlaybackSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [titleId]
              properties:
                titleId:
                  type: string
                assetId:
                  type: string
                profileId:
                  type: string
                  default: local-profile
                deviceId:
                  type: string
                  default: local-browser
                territory:
                  type: string
                  default: NG
                  minLength: 2
                  maxLength: 2
      responses:
        '200':
          description: Authorized playback source
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackSource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  responses:
    Problem:
      description: RFC 7807-like problem response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
    Artwork:
      type: object
      required: [url, kind, alt, width, height, dominantColor]
      properties:
        url:
          type: string
        kind:
          enum: [poster, landscape, hero, logo]
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
        dominantColor:
          type: string
    TitleSummary:
      type: object
      required: [id, slug, kind, title, synopsis, releaseYear, maturityRating, genres, countries, artwork, availableFrom, availableUntil, minimumPlan, offlineDownloadAllowed]
      properties:
        id:
          type: string
        slug:
          type: string
        kind:
          enum: [movie, series, episode, trailer]
        title:
          type: string
        synopsis:
          type: string
        releaseYear:
          type: integer
        runtimeSeconds:
          type: [integer, 'null']
        maturityRating:
          enum: [G, PG, PG_13, R, NC_17]
        genres:
          type: array
          items:
            type: string
        countries:
          type: array
          items:
            type: string
        artwork:
          type: array
          items:
            $ref: '#/components/schemas/Artwork'
        availableFrom:
          type: string
          format: date-time
        availableUntil:
          type: string
          format: date-time
        minimumPlan:
          enum: [guest, free, standard, premium, admin]
        offlineDownloadAllowed:
          type: boolean
...
    Track:
      type: object
      required: [language, label, kind, accessibility]
      properties:
        language:
          type: string
        label:
          type: string
        kind:
          enum: [subtitles, captions, audio]
        accessibility:
          type: boolean
...
    SearchResponse:
      type: object
      required: [query, results, suggestions]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/TitleSummary'
        suggestions:
          type: array
          items:
            type: string
    PlaybackSource:
      type: object
      required: [sessionId, playbackUrl, sourceType, expiresAt, drmActive, drmScheme, captions, audio, heartbeatIntervalSeconds, policy]
      properties:
        sessionId:
          type: string
        playbackUrl:
          type: string
        sourceType:
          enum: [mp4, hls, dash]
        expiresAt:
          type: string
          format: date-time
        drmActive:
          type: boolean
        drmScheme:
          enum: [none, widevine, fairplay, playready]
        captions:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        audio:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        heartbeatIntervalSeconds:
          type: integer
        policy:
          type: object
...